fix: fixes for gameoflive

Created-by: liontiger23
Commit-by: Ivan Trepakov
Merged-by: liontiger23
Description: Various fixes for issues discovered with gameoflive benchmark including:
- SccImm parsing
- Logging improvement (correct jump offsets and register value logging in ectype->Put)
- MemTailStore argument order
- MemBodyIndex support
- Throw instruction stub support

See merge request: wellox/cbc-engine-initial-stage!123

+++ src/interpreter/interpreter.h
@@ -295,21 +295,21 @@ public:
 
     template <Width::Value width> inline void SCC(CC cc, IReg d, IReg l, IReg r)
     {
-        uint32_t res = Cmp<width>(cc, l, r) ? 1 : 0;
-        ectype->Put(d, Value::Primitive { .u32 = res });
+        uint64_t res = Cmp<width>(cc, l, r) ? 1 : 0;
+        ectype->Put(d, Value::Primitive { .u64 = res });
     }
 
     template <Width::Value width> inline void SCC(CC cc, IReg d, FReg l, FReg r)
     {
-        uint32_t res = Cmp<width>(cc, l, r) ? 1 : 0;
-        ectype->Put(d, Value::Primitive { .u32 = res });
+        uint64_t res = Cmp<width>(cc, l, r) ? 1 : 0;
+        ectype->Put(d, Value::Primitive { .u64 = res });
     }
 
     template <RT::ImmKind::Value immKind, Width::Value width> inline void SCCImm(CC cc, IReg d, IReg l, uint16_t imm)
     {
-        uint32_t res =
+        uint64_t res =
             CmpPrim<width>(cc, l, Value::Primitive { .u64 = DecodeImmediate<immKind>(literals, imm) }) ? 1 : 0;
-        ectype->Put(d, Value::Primitive { .u32 = res });
+        ectype->Put(d, Value::Primitive { .u64 = res });
     }
 
     void Convert(ConvertType toType, ConvertType fromType, Reg to, Reg from)
